src/cdc_producer/requirements.txt
python-dotenv>=1.0.0
pyarrow>=22.0.0
pandas>=2.3.3
numpy>=2.2.6
psycopg==3.3.2
sqlalchemy==2.0.45
alembic==1.17.2
fastparquet==2025.12.0
psycopg2-binary==2.9.11